requirements.txt
supabase
python-dotenv
pandas
geopandas
rasterio
sentinelhub
numpy